TextsToolsServer: рефакторинг

+++ TextsTool/TextsToolServer/DbSerializer.cpp
@@ -110,13 +110,13 @@ void DbSerializer::SaveDatabase()
 	buffer.PushStringWithoutZero("TDBF0001");
 
 	buffer.Push(_pDataBase->_newAttributeId);
-	buffer.Push(_pDataBase->_attributeProps.size());
+	buffer.Push((uint32_t)_pDataBase->_attributeProps.size());
 	for (const auto& atribProp : _pDataBase->_attributeProps) {
 		atribProp.SaveToBase(buffer);
 	}
 
 	buffer.Push(_pDataBase->_newFolderId);
-	buffer.Push(_pDataBase->_folders.size());
+	buffer.Push((uint32_t)_pDataBase->_folders.size());
 	for (const auto& folder : _pDataBase->_folders) {
 		folder.SaveToBase(buffer);
 	}
